Cleaned Up Code
Made a Seperate folder for the test code. I have tested the interrupt and figured out that if it is too fast, the code breaks. So I slowed it down. The math functions also work. Only problem now is the input from the controller. I also created a barebones sketch to test out the problems with the controller. Cause the example works, but my implementation does not.