Speed pot implemented
I have implemented the speed control potentionmeter. The drift control pot would not hold a stable value. So while the code works, I have disabled that for now. I have also cleaned and commented the code. Next steps are to test the drift pot with a multi-meter so that I can make sure it is the problem.